HTCONDOR-2753: Better file and function naming
