+++ src/condor_starter.V6.1/starter.cpp
@@ -56,7 +56,7 @@
 #include "data_reuse.h"
 #endif
 #include "authentication.h"
-#include "format_bytes.h"
+#include "to_string_si_units.h"
 
 extern void main_shutdown_fast();
 
@@ -4117,7 +4117,7 @@ Starter::CheckLVUsage( int /* timerID */ )
 
 	if (monitor->du.execute_size >= limit) {
 		std::string hold_msg;
-		std::string limit_str = format_bytes(limit);
+		std::string limit_str = to_string_byte_units(limit);
 		formatstr(hold_msg, "Job has exceeded allocated disk (%s). Consider increasing the value of request_disk.",
 		         limit_str.c_str());
 		jic->holdJob(hold_msg.c_str(), CONDOR_HOLD_CODE::JobOutOfResources, 0);
