+++ src/condor_utils/CMakeLists.txt
@@ -301,8 +301,6 @@ forkwork.cpp
 forkwork.h
 format_time.cpp
 format_time.h
-format_bytes.cpp
-format_bytes.h
 fs_util.cpp
 gahp_common.cpp
 gahp_common.h
@@ -464,6 +462,8 @@ tokener.cpp
 tokener.h
 token_utils.cpp
 token_utils.h
+to_string_si_units.cpp
+to_string_si_units.h
 translation_utils.cpp
 translation_utils.h
 tree.h
